pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "city2graph"
version = "0.1.0"
description = "Transform geospatial relations into graphs for spatial analysis and GNNs."
authors = [
    {name = "Yuta Sato", email = "y.sato@liverpool.ac.uk"}
]
readme = "README.md"
license = {text = "BSD-3-Clause"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Scientific/Engineering :: GIS",
]
keywords = ["graph-neural-networks", "pytorch-geometric", "urban", "transportation", "mobility", "overturemaps", "gtfs"]
requires-python = ">=3.11,<4.0"
dependencies = [
    "networkx >=2.8",
    "osmnx >=2.0.3",
    "shapely >=2.1.0",
    "geopandas >0.12.0",
    "libpysal >=4.12.1",
    "momepy",
    "overturemaps",
]

[project.optional-dependencies]
torch = [
    "torch >=2.4.0",
    "torch_geometric >=2.6.1",
    # PyG extensions need to be installed separately
]

[dependency-groups]
docs = [
    "sphinx",
    "sphinx-autodoc-typehints",
    "pydata-sphinx-theme",
    "myst-parser",
    "nbsphinx",
    "numpydoc",
    "sphinxext-opengraph[social-cards] >=0.10.0",
    "sphinx-plotly-directive",
    "sphinx-sitemap",
    "toml",
    "pandoc",
    "overturemaps"
]
dev = [
    "pytest",
    "pytest-cov",
    "ipython >=8.14.0",
    "jupyter >=1.0.0",
    "notebook >=7.0.0",
    "isort >=5.12.0",
    "ipykernel >=6.29.5",
    "matplotlib >=3.10.3",
    "folium >=0.19.6",
    "mapclassify >=2.8.1",
    "contextily>=1.6.2",
]

[project.urls]
"Homepage" = "https://github.com/c2g-dev/city2graph"
"Documentation" = "https://city2graph.net"
"Bug Tracker" = "https://github.com/c2g-dev/city2graph/issues"

[tool.uv]
default-groups = []

[tool.hatch.build.targets.wheel]
packages = ["city2graph"]

[tool.ruff]
cache-dir = "~/.cache/pre-commit/ruff"
exclude = [
    "docs/*",
    "dev/*",
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

line-length = 100

[tool.ruff.lint]
extend-ignore = [
    "DTZ001",
    "DTZ007",
    "FBT001",
    "FBT002",
    "N803",
    "N806",
    "PLR2004",
    "SLF001",
    "S101",
    "S603",
    "TRY300",
    ]
extend-select = ["ALL"]

[tool.ruff.lint.isort]
force-single-line = true

[tool.ruff.lint.mccabe]
max-complexity = 14

[tool.ruff.lint.pycodestyle]
max-line-length = 110 # line length + 10% since it isn't a hard upper bound

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.pylint]
max-args = 10
max-branches = 14